Add BLE command parser & refactor configuration into class.
This commit adds BLE functionality and a new command parser for use with
the Companion App. The other major change is that instead of having a
global configuration struct, it's been refactored into a class (with
methods for manipulating the configuration).

This also includes some reference docs of the commands useable over
Serial & BLE.