Initial incomplete incorporation of timed sequence functions
Timed sequences use arrays of functions, times, and sequences to step through actions over specified times and orders. Added AsyncDelay library for timing instead of millis(). Goal is to create adaptable rinse and prime routines. Functions need to be evaluated for completeness. Logic is not finished for how to exit time sequence in main code.